requirements.txt

pytest
CONFIG
requests
yaml
json
allure-pytest
os
base64
